requirements.txt
pypresence
pytrakt
python-dotenv
